requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
alabaster==0.7.13
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astor @ file:///home/conda/feedstock_root/build_artifacts/astor_1593610464257/work
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
Babel==2.15.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
bounded-pool-executor==0.0.3
braceexpand==0.1.7
Brotli @ file:///C:/Windows/Temp/abs_63l7912z0e/croots/recipe/brotli-split_1659616056886/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1707022139797/work/certifi
cffi==1.16.0
chardet==3.0.4
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
clip==0.2.0
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.2
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.1
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
defusedxml==0.7.1
Distance==0.1.3
dnspython==2.6.1
docutils==0.20.1
docx2pdf==0.1.8
easydict==1.13
email_validator==2.1.1
et-xmlfile==1.1.0
exceptiongroup==1.2.1
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.3
fastjsonschema==2.19.1
ffmpy==0.3.2
filelock @ file:///C:/b/abs_f2gie28u58/croot/filelock_1700591233643/work
fire==0.6.0
flatbuffers==24.3.25
fonttools==4.51.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.3.1
gmpy2 @ file:///C:/ci/gmpy2_1645456279018/work
google-trans-new==1.1.9
googletrans==3.0.0
gradio==4.31.2
gradio_client==0.16.3
graphviz==0.20.3
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.5.1
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.0
hyperframe==5.2.0
idna==2.10
imagesize==1.4.1
imbalanced-learn==0.12.2
imblearn==0.0
importlib_metadata==7.1.0
importlib_resources==6.4.0
iniconfig==2.0.0
install==1.3.5
ipykernel==6.29.4
ipython==8.12.3
ipywidgets==8.1.2
isoduration==20.11.0
jedi==0.19.1
jieba==0.42.1
Jinja2 @ file:///C:/b/abs_f7x5a8op2h/croot/jinja2_1706733672594/work
joblib==1.4.2
json5==0.9.25
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.8
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.1
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
llvmlite==0.41.1
lmdb==1.4.1
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe @ file:///C:/b/abs_ecfdqh67b_/croot/markupsafe_1704206030535/work
matplotlib==3.7.5
matplotlib-inline==0.1.7
mdurl==0.1.2
missingno==0.5.2
mistune==3.0.2
mkl-fft @ file:///C:/b/abs_19i1y8ykas/croot/mkl_fft_1695058226480/work
mkl-random @ file:///C:/b/abs_edwkj1_o69/croot/mkl_random_1695059866750/work
mkl-service==2.4.0
mpmath @ file:///C:/b/abs_7833jrbiox/croot/mpmath_1690848321154/work
multidict==6.0.5
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx @ file:///C:/b/abs_e6gi1go5op/croot/networkx_1690562046966/work
notebook==7.1.3
notebook_shim==0.2.4
numba==0.58.1
numpy @ file:///C:/Users/dev-admin/mkl/numpy_and_numpy_base_1682982345978/work
numpydoc==1.7.0
openai==0.28.0
OpenCC==1.1.7
opencv-python-headless==4.9.0.80
openpyxl==3.1.2
opt-einsum @ file:///home/conda/feedstock_root/build_artifacts/opt_einsum_1696448916724/work
orjson==3.10.3
overrides==7.7.0
packaging==24.0
paddle-bfloat==0.1.7
paddlepaddle-gpu==2.4.0
paddlespeech==0.0.0
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.4
pattern-singleton==1.2.0
pdf2docx==0.5.8
PDPbox==0.3.0
PDPbox1==0.2.5
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.1
plotly==5.22.0
pluggy==1.5.0
portalocker==2.8.2
pox==0.3.4
ppft==1.7.6.8
pqdm==0.2.0
praatio==5.0.0
prettytable==3.10.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf @ file:///D:/bld/protobuf_1664075564769/work/python
psutil==5.9.8
pure-eval==0.2.2
PyAudio==0.2.14
pybind11==2.12.0
pyclipper==1.3.0.post5
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
pydub==0.25.1
PyExecJS==1.5.1
Pygments==2.18.0
PyMuPDF==1.20.2
PyMuPDFb==1.24.3
pynput==1.7.7
pyparsing==3.1.2
PyPDF2==3.0.1
pyperclip==1.8.2
pypinyin==0.44.0
pypinyin-dict==0.8.0
PySocks @ file:///C:/ci/pysocks_1605287845585/work
pysrt==1.1.2
pystray==0.19.5
pytest==8.2.0
pytest-runner==6.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.9
python-pptx==0.6.23
pytz==2024.1
PyWavelets==1.4.1
pywin32==306
pywinpty==2.0.13
PyYAML @ file:///C:/b/abs_782o3mbw7z/croot/pyyaml_1698096085010/work
pyzmq==26.0.3
qtconsole==5.5.2
QtPy==2.4.1
rapidfuzz==3.9.0
rarfile==4.2
referencing==0.35.1
regex==2024.5.10
reportlab==4.2.0
requests @ file:///C:/b/abs_474vaa3x9e/croot/requests_1707355619957/work
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.1
ruff==0.4.4
safetensors==0.4.3
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.3
sentencepiece==0.1.99
shap==0.44.1
shapely==2.0.4
shellingham==1.5.4
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
slicer==0.0.7
sniffio==1.3.1
snowballstemmer==2.2.0
some-package==0.1
soupsieve==2.5
Sphinx==7.1.2
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
stack-data==0.6.3
starlette==0.37.2
sympy @ file:///C:/b/abs_82njkonm7f/croot/sympy_1701397685028/work
tabulate==0.9.0
tenacity==8.3.0
termcolor==2.4.0
terminado==0.18.1
TextGrid==1.6.1
threadpoolctl==3.5.0
tifffile==2023.7.10
timer==0.2.2
tinycss2==1.3.0
tokenizers==0.19.1
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.2.2
torchaudio==2.2.2
torchkeras==3.9.6
torchvision==0.17.2
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
transformers==4.40.2
ttkthemes==3.2.2
typer==0.12.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.10.0
uri-template==1.3.0
urllib3 @ file:///C:/b/abs_4etpfrkumr/croot/urllib3_1707770616184/work
uvicorn==0.29.0
watchfiles==0.21.0
Wave==0.0.2
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
webrtcvad==2.0.10
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==3.0.3
widgetsnbextension==4.0.10
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306167264/work
win32-setctime==1.1.0
xgboost==2.0.3
XlsxWriter==3.2.0
yacs==0.1.8
yarl==1.9.4
zhon==2.0.2
zipp==3.18.1
